+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+### Added
+
+- Support Curve add liquidity in `HCurve`.
+- Support Curve remove liquidity one coin in `HCurve`.
+
 ## [0.8.0] - 2020-08-10
 
 ### Added
@@ -15,8 +20,6 @@ All notable changes to this project will be documented in this file.
 - Support OneInch fixed input eth to token swap.
 - Support OneInch fixed input token to eth swap.
 - Support OneInch fixed input token to token swap.
-- Support Curve add liquidity in `HCurve`.
-- Support Curve remove liquidity one coin in `HCurve`.
 
 ### Changed