+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - Support Curve add liquidity in `HCurve`.
 - Support Curve remove liquidity one coin in `HCurve`.
 - Support WETH deposit and withdraw function in `HWeth`.
+- Support Curve deposit to gauges in `HCurveDao`.
+- Support Curve mint and mint_many CRV in `HCurveDao`.
 
 ## [0.8.0] - 2020-08-10